backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Redis
redis[hiredis]==5.2.1

# Validation
email-validator==2.2.0

# Auth
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# Utilities
python-multipart==0.0.20
